requirements.txt
Fabric3==1.13.1.post1
dnspython==1.15.0
Jinja2==2.10
